add test coverage for the case fixed by 764c1afde6998fd86d4e0d8b5e561402de94a2a9
also fixed another bug where `NoSuchMethodError` was being constructed
with a String instead of a Symbol - 4 of the tests were failing on my
machine (Windows 7) prior to this.  I implemented a fix which passes
both the String and Symbol to `_forward`, so as to not require a call to
`new Symbol` from inside the `_forward` method; it looks ugly but is
more efficient.